Move MatrixStore fixture reuse code into one place
